Updating setup_linux_pescar.sh to support Mint 18

+++ setup_linux_pescar.sh
@@ -19,6 +19,7 @@
 # - Update/Install some packages
 #
 #              CHANGE HISTORY
+# 3.2 (12 Jun 2017) - Improving log messages and updating script to support Mint 18
 # 3.1.1 (05 Jun 2016) - Removed Proxy information (in order to public the source)
 # 3.1 (05 Jun 2016) - Added again the functions to set proxy to Firefox and Chormium. Added Zenity to display the options sbox
 # 3.0 (25 May 2016) - Refectored all functions. Removed functions to set proxy to Firefox and Chormium
@@ -28,7 +29,6 @@
 # 1.1 (14 May 2015) - Added dialog as default UI
 # 1.0 (08 Jun 2014) - First version
 #
-
 
 # Setting up the global variables
 
@@ -80,7 +80,7 @@ APT_GET_UPDATE=1
 #   None
 #######################################
 function logger(){
-    echo -en "${*}\n" | tee -a "${TMP_LOG_FILE}"
+    echo -en "$(date) - ${FUNCNAME[1]}(): ${*}\n" | tee -a "${TMP_LOG_FILE}"
 }
 
 
@@ -104,7 +104,7 @@ function print_header(){
 
     header="\n+${header}+\n| ${phrase} |\n+${header}+\n\n"
 
-    logger "${header}"
+    echo -en "${header}"
 }
 
 
@@ -118,7 +118,7 @@ function print_header(){
 #   None
 #######################################
 function print_line(){
-    logger '----------------------------------------------------------------------'
+    echo '----------------------------------------------------------------------'
 }
 
 
@@ -137,14 +137,14 @@ function backup(){
 	local backup_name="${file}.bkp.$(date +%s)"
     local exit_code=2
 
-    print_header "Creating backup for ${file}"
+    logger "Creating backup for ${file}"
 
     if [[ -f "${file}" ]]; then
-        cp -fv "${file}" "${backup_name}" | tee -a "${TMP_LOG_FILE}"
+        cp -f "${file}" "${backup_name}" | tee -a "${TMP_LOG_FILE}"
         exit_code="$?"
 
     elif [[ -d "${file}" ]]; then
-        cp -Rfv "${file}" "${backup_name}" | tee -a "${TMP_LOG_FILE}"
+        cp -Rf "${file}" "${backup_name}" | tee -a "${TMP_LOG_FILE}"
         exit_code="$?"
 
     else
@@ -248,14 +248,33 @@ function run_dpkg_all(){
 function configure_firefox_profile(){
     local tar_file='firefox_profile.tar.gz'
     local pescar_profile='/home/pescar'
+    local firefox_profile_basedir="${pescar_profile}/.mozilla/firefox"
+    local default_pescar_profile_dir='default_pescar_profile'
+    local current_firefox_profile_name=''
+    
+    cd "${BASE_DIR}"
 
-    backup "${pescar_profile}/.mozilla"
+    print_header 'Configure Firefox profile'
 
-    cp -vf  "${BASE_DIR}/${tar_file}" "${pescar_profile}"
-    cd "${pescar_profile}"
-    tar -zxvf "${tar_file}"
-    chown -R pescar.pescar "${pescar_profile}/.mozilla"
-    rm -f "${pescar_profile}/${tar_file}"
+    if ls -d ${firefox_profile_basedir}/*.default 2> /dev/null; then
+        current_firefox_profile_name=$(basename $(ls -d ${firefox_profile_basedir}/*.default))
+    else
+        logger "Unable to find any profile folder under \"${firefox_profile_basedir}\". Skipping profile changes."
+        return 1
+    fi
+
+    backup "${firefox_profile_basedir}"
+
+    if cp -f  "${BASE_DIR}/${tar_file}" "${firefox_profile_basedir}"; then
+        cd "${firefox_profile_basedir}"
+        tar -zxf "${tar_file}"
+        rm -rf "${current_firefox_profile_name}"
+        
+        mv "${default_pescar_profile_dir}" "${current_firefox_profile_name}"
+        chown -R pescar.pescar "${current_firefox_profile_name}"
+        rm -f "${tar_file}"
+    fi
+
 }
 
 
@@ -268,21 +287,44 @@ function configure_firefox_profile(){
 # Returns:
 #   None
 #######################################
-function update_system(){
+function update_system() {
+    local run_apt_update=false
+    local run_dpkg=false
     local apt_target_basedir='/var/cache'
     local tar_file='apt.tar'
+
+    if grep -iq 'run-update' <<< "$@"; then
+        run_apt_update=true
+    fi
+
+    if grep -iq 'run-dpkg' <<< "$@"; then
+        run_dpkg=true
+    fi
 
     print_header "Updating the system"
 
     backup "${apt_target_basedir}/apt"
     rm -rf "${apt_target_basedir}/apt"
-    cp -f "${BASE_DIR}/${tar_file}" "${apt_target_basedir}"
-    cd "${apt_target_basedir}"
-    tar -xvf "${tar_file}"
 
-    run_apt_get_update
-    run_dpkg_all
+    logger "Copying installation files from \"${BASE_DIR}/${tar_file}\" to \"${apt_target_basedir}\"" 
+    cp -f "${BASE_DIR}/${tar_file}" "${apt_target_basedir}"
+    logger 'Copy done.'
+
+    cd "${apt_target_basedir}"
+    logger "Extracting file \"${tar_file}\""
+    tar -xf "${tar_file}"
+    logger 'Done!'
+
+    if ${run_dpkg}; then
+        run_dpkg_all
+    fi
+
+    if ${run_apt_update}; then
+       run_apt_get_update       
+    fi
+ 
     configure_firefox_profile
+ 
     add_root_to_ssh
 
     print_line
@@ -301,6 +343,11 @@ function update_system(){
 #######################################
 function set_proxy_to_file(){
     local dest_file="${1}"
+
+    if [[ ! -f "${dest_file}" ]]; then
+        logger "set_proxy_to_file(): Sent at least one file as parameter"
+        return 1
+    fi
 
     print_header "Setting proxy to file: ${dest_file}"
 
@@ -332,9 +379,15 @@ function set_proxy_to_chromium(){
   local chromium_bin="${chromium}-bin"
 
   print_header 'Set proxy to Chromium'
+    
+  if [[ ! -f "${chromium}" ]]; then
+    logger "Chromiun Browser not found. Skipping proxy configuration."
+    return 1
+  fi
 
+  backup "${chromium}"
   mv "${chromium}" "${chromium_bin}"
-
+  
   {
       echo "#!/bin/bash"
       echo -e "\n\n"
@@ -365,6 +418,8 @@ function set_proxy_to_chromium(){
 #######################################
 set_proxy_to_firefox(){
   local firefox="$(which firefox)"
+  print_header 'Set proxy to Firefox'
+  backup "${firefox}"
 
   sed -i "2i# Proxy Configuration\n\
   export http_proxy='${PROXY}'\n\
@@ -517,8 +572,8 @@ function start_set_up(){
             --column "${DIALOG_OPTIONS}" \
             TRUE "${DIALOG_OPT_PROXY_SYSTEM}" \
             TRUE "${DIALOG_OPT_PROXY_APT}" \
-            TRUE "${DIALOG_OPT_UPDATE}" \
-            TRUE "${DIALOG_OPT_BROWSER_PROXY}")
+            TRUE "${DIALOG_OPT_BROWSER_PROXY}" \
+            FALSE "${DIALOG_OPT_UPDATE}")
 
     options="$(tr '|' ' ' <<< ${options})"
 
@@ -536,26 +591,18 @@ function start_set_up(){
                 set_proxy_to_apt
            ;;
            "${DIALOG_OPT_UPDATE}")
-                logger "${MESSAGE_UPDATING_CONFIG}"
-                update_system
+                update_system --run-dpkg #--run-update
            ;;
            "${DIALOG_OPT_BROWSER_PROXY}")
                 set_proxy_to_firefox
                 set_proxy_to_chromium
+                configure_firefox_profile
            ;;
            *)
                 logger "Option ${opt} not found! exiting."
            ;;
        esac
    done
-
-
-    if [[ -f "${TMP_LOG_FILE}" ]]; then
-        logger "${MESSAGE_OPEN_LOG_FILE}"
-        nohup gedit "${TMP_LOG_FILE}" & 2> /dev/null
-        [[ -f 'nohup.out' ]] && rm -f 'nohup.out'
-    fi
-
 
 }
 
@@ -575,16 +622,21 @@ function start_set_up(){
 # Returns:
 #   None
 #######################################
-function main(){
+function main(){ 
+    local time_to_wait=30
 
     if [[ "$(id -u)" == "0" ]]; then
         # If we are here, the script is being run as root.
         touch "${TMP_LOG_FILE}"
         chmod 777 "${TMP_LOG_FILE}"
+
         start_set_up
+
         touch "${IT_HAS_RUN}"
-        logger 'Exiting in 10 seconds.'
-        sleep 10
+        chown pescar.pescar "${IT_HAS_RUN}"
+
+        logger "Exiting in ${time_to_wait} seconds."
+        sleep ${time_to_wait}
     else
         # Try to use a default password to run this script
         sudo -S "${THIS_SCRIPT_PATH}" <<< "${PESCAR_DEFAULT_PASSWORD}"
@@ -593,12 +645,10 @@ function main(){
         if [[ ! -f "${IT_HAS_RUN}" ]]; then
             clear
             print_header "${MESSAGE_ENTER_PASSWORD} ${USER}"
-            sudo "${THIS_SCRIPT_PATH}"
         fi
 
         [[ -f "${IT_HAS_RUN}" ]] && rm -f "${IT_HAS_RUN}"
     fi
-
 
 }
